+++ astrodynamics/cubesat_analysis_tools/mag_field.py
